+++ firmware/esp32_edge_impulse/components/eye_ui/eye_ui.h
@@ -4,6 +4,9 @@
 extern "C" {
 #endif
 
+// Display power control pin
+#define DISPLAY_EN_PIN   GPIO_NUM_4   // GPIO to control display power
+
 // Call initArduino() in app_main() before calling this.
 void eye_ui_start(void);